+++ src/backend/node/commands/TOWARDS.java
@@ -21,7 +21,7 @@ public class TOWARDS extends TwoArgumentNode {
 		int targetY = getAndRun(1, mc).getIntegerValue();
 
 		ITurtleLambda l = (Turtle t) -> {
-			int[] currPos = t.getPosition();
+			double[] currPos = t.getPosition();
 			double theta = Math.atan2(-targetY - currPos[1], targetX - currPos[0]);
 			double angle = Math.toDegrees(theta);
 			angle = 360 - angle;
